% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{exp2lat}
\alias{exp2lat}
\title{exp2lat}
\usage{
exp2lat(mean, cov)
}
\arguments{
\item{mean}{a vector of means. These need to all be >0.}

\item{cov}{a single variance or a covariance matrix}
}
\value{
a list, which an element 'mean' for the transformed mean(s) and another element 'cov' for the transformed covariance matrix
}
\description{
Transforms means and covariance matrix from expected (log-normal) to latent (normal) scale
}
\details{
Transforms expected mean(s) and (co)variance(s) of a multivariate normal distribution to expected mean(s) and (co)variance(s) of a multivariate log-normal distribution. This is useful when trying to go from expected (log-normal) to expected (log-normal) scales in a Poisson GLM with a log link function.
}
\examples{
exp2lat(1,1)

}
\author{
Joel Pick - joel.l.pick@gmail.com
}
