% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_variances.R
\name{expected_variance}
\alias{expected_variance}
\title{expected_variance}
\usage{
expected_variance(squid)
}
\arguments{
\item{squid}{A squid object}
}
\value{
A data.frame with the data structure
}
\description{
Calculate expected variance in response variable(s)
}
\details{
Calculated the expect variance from the simulation parameters. Has several limitaitons. 
1. Doesn't work when 'model' is specific in simulate_population
2. Assumes random factors are balanced - unbalanced designs will change the observed variance
3. Will be inaccurate with transformed variables (i.e. if functions are specified)
4. Doesn't deal well with three way interactions and over, unless they are the same variable (i.e. polynomials)
5. Doesn't account for covariance between interaction terms, e.g. if there is are two interaction terms rain:temp and wind:temp, they will covary, but this additional variance in the response is not calculated
}
\examples{
# simple data structure with 5 'individuals' and 2 observations per individual
make_structure(structure="individual(5)", repeat_obs=2)

# nested data structure with 2 sexes, 5 individuals per sex
# and 2 observations per individual
make_structure(structure="sex(2)/individual(5)", repeat_obs=2)

# crossed data structure with 5 individuals in 2 treatments 
# and 2 observations per individual and treatment combination
make_structure(structure="treatment(2) + individual(5)", repeat_obs=1)

}
\author{
Joel Pick - joel.l.pick@gmail.com
}
