% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{lat2exp}
\alias{lat2exp}
\title{lat2exp}
\usage{
lat2exp(mean, cov)
}
\arguments{
\item{mean}{a vector of means}

\item{cov}{a single variance or a covariance matrix}
}
\value{
a list, which an element 'mean' for the transformed mean(s) and another element 'cov' for the transformed covariance matrix
}
\description{
Transforms means and covariance matrix from latent (normal) to expected (log-normal) scale
}
\details{
Transforms expected mean(s) and (co)variance(s) of a multivariate log-normal distribution to expected mean(s) and (co)variance(s) of a multivariate normal distribution. This is useful when trying to go from latent (normal) to expected (log-normal) scales in a Poisson GLM with a log link function.
}
\examples{
lat2exp(1,1)

}
\author{
Joel Pick - joel.l.pick@gmail.com
}
