% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expected_variances.R
\name{print.squid_var}
\alias{print.squid_var}
\title{print.squid_var}
\usage{
\method{print}{squid_var}(x, ...)
}
\arguments{
\item{x}{an R object of class 'squid_var'}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print method for class 'squid_var'
}
