% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_population.R
\name{simulate_population}
\alias{simulate_population}
\title{simulate_population}
\usage{
simulate_population(
  data_structure,
  N,
  parameters,
  N_response = 1,
  known_predictors,
  model,
  family = "gaussian",
  link = "identity",
  pedigree,
  pedigree_type,
  phylogeny,
  phylogeny_type,
  cov_str,
  sample_type,
  sample_param,
  sample_plot = FALSE,
  N_pop = 1
)
}
\arguments{
\item{data_structure}{A matrix or dataframe with a named column for each grouping factor, including the levels}

\item{N}{Sample size when data_structure is not specified}

\item{parameters}{A list of parameters for each hierarchical level. See details.}

\item{N_response}{The number of response variables, defaults to 1.}

\item{known_predictors}{This option provides a way of inputting existing predictor variables, without simulating all predictors. This argument takes a list, with item 'predictors' and 'betas', where 'predictors' is a matrix or dataframe}

\item{model}{Optional.}

\item{family}{A description of the error distribution. Default "gaussian".}

\item{link}{A description of the link function distribution. Default "identity".}

\item{pedigree}{A list of pedigrees for each hierarchical level. Each pedigree must be matrix or data.frame, that is at least 3 columns, which correspond to ID, dam and sire.}

\item{pedigree_type}{A list describing what kind of genetic variance is to be simulated from each pedigree. Default is 'additive', other options are 'dominance' and 'epistatic'. Makes use of relationship matrices created by the nadiv package.}

\item{phylogeny}{A list of phylogenies for each hierarchical level. Each pedigree should be phylo class.}

\item{phylogeny_type}{A list describing what mode of evolution should be simulated from each phylogeny. Options are 'brownian'(default) or 'OU'.}

\item{cov_str}{A list of covariance structures for each hierarchical level.}

\item{sample_type}{Type of sampling, needs to be one of 'nested', 'missing' or temporal. See details}

\item{sample_param}{A set of parameters, specific to the sampling type. See details.}

\item{sample_plot}{Logical. Should illustrative plots be made - defaults to FALSE.}

\item{N_pop}{Number of populations. Default = 1}
}
\value{
a squid object, which is a list including all inputs and simulated data.
}
\description{
Simulate population level data
}
\details{
Parameter list ...
}
\examples{
# simple linear model with three predictors variables
squid_data <- simulate_population(
  parameters = list(
    observation = list(
      names = c("temperature","rainfall", "wind"),
      beta = c(0.5,-0.3, 0.4)    
  ),
    residual = list(
      vcov = 1
    )
  ),
  N=2000
)

}
\author{
Joel Pick - joel.l.pick@gmail.com
}
