% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_structure.R
\name{make_structure}
\alias{make_structure}
\title{make_structure}
\usage{
make_structure(structure, repeat_obs = 1, ...)
}
\arguments{
\item{structure}{A formula specifying the structure and sample sizes at each level. See details.}

\item{repeat_obs}{Number of repeated observations at the lowest level}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
A data.frame with the data structure
}
\description{
Make nested and crossed balanced hierarchical structures
}
\details{
Factors are input as a text string. The name of each factor is followed by the number of levels in that factor in brackets e.g. "individual(100)". Nested factors can be specified using "/", e.g. "population(2)/individual(2)", the lower levels being specified after the higher levels, and the sample sizes of the lower levels
}
\examples{
# simple data structure with 5 'individuals' and 2 observations per individual
make_structure(structure="individual(5)", repeat_obs=2)

# nested data structure with 2 sexes, 5 individuals per sex
# and 2 observations per individual
make_structure(structure="sex(2)/individual(5)", repeat_obs=2)

# crossed data structure with 5 individuals in 2 treatments 
# and 2 observations per individual and treatment combination
make_structure(structure="treatment(2) + individual(5)", repeat_obs=1)

}
\author{
Joel Pick - joel.l.pick@gmail.com
}
